backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.8.2
python-multipart==0.0.6
httpx==0.25.2
python-dateutil==2.8.2
pytz==2023.3
email-validator==2.1.0
jinja2==3.1.2
aiofiles==23.2.1
